requirements.txt
anthropic>=0.25.0
feedparser>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
notion-client>=2.2.0
